stripeservice/requirements.txt
attrs==19.3.0
awscli==1.18.125
boto3==1.14.48
botocore==1.17.48
certifi==2020.6.20
cffi==1.14.2
chalice==1.18.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
docutils==0.15.2
enum-compat==0.0.3
idna==2.10
jmespath==0.10.0
mypy-extensions==0.4.3
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.5
s3transfer==0.3.3
six==1.15.0
stripe==2.50.0
urllib3==1.25.10
